java

+++ Java_Programs/Area_of_circle.java
@@ -1,7 +1,7 @@
 //Calculate the area of a circle given its radius using the formula: A = π * r^2.
 public class Area_of_circle {
     public static void main(String[] args) {
-        int radius = 5; // Example radius
+        double radius = 5.0; // Radius of the circle
         final double pi = 3.14; //  Constant value of π
         double area = pi * radius * radius; // Calculate area
         System.out.println("The area of circle is : " + area); // Printing the area of circle
